backend/requirements.txt
# pip install -r requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
Flask-Migrate